solved NA's in subscripts problem in MCMC
by dropping all the NA's from the update objects. Furthermore, I
simulated the data again setting all males to be migrators (not only 70
percent) and made no rules for who has an observed death date in the
simulated data. Ran the model on the simulated data. It shows that we
are still not retrieving the male parameters correctly.